fix(on deck hooks): refactored position calculations + updater + erro handling + optimistic
