+++ app/client/src/features/events/ModeratorView/hooks/OnDeck/useOnDeckDequeued.ts
@@ -5,6 +5,9 @@ import { GraphQLSubscriptionConfig } from 'relay-runtime';
 import type { useOnDeckDequeuedSubscription } from '@local/__generated__/useOnDeckDequeuedSubscription.graphql';
 import { useEvent } from '@local/features/events/useEvent';
 import { Question, Topic } from '../../types';
+import { calculateOnDeckDequeuePosition } from './utils';
+import { useOnDeckDequeuedMutation } from '@local/__generated__/useOnDeckDequeuedMutation.graphql';
+import { useSnack } from '@local/core';
 
 export const USE_ON_DECK_DEQUEUED_MUTATION = graphql`
     mutation useOnDeckDequeuedMutation($input: RemoveQuestionFromOnDeck!) {
@@ -17,12 +20,7 @@ export const USE_ON_DECK_DEQUEUED_MUTATION = graphql`
                     id
                     position
                     onDeckPosition
-                    ...QuestionAuthorFragment
-                    ...QuestionStatsFragment
-                    ...QuestionContentFragment
                     topics {
-                        topic
-                        description
                         position
                     }
                 }
@@ -37,11 +35,6 @@ export const USE_ON_DECK_DEQUEUED = graphql`
             edge {
                 node {
                     id @deleteEdge(connections: $connections)
-                    position
-                    onDeckPosition
-                    topics {
-                        position
-                    }
                 }
                 cursor
             }
@@ -57,7 +50,9 @@ interface Props {
 
 export function useOnDeckDequeued({ connections, topic: currentTopic }: Props) {
     const { eventId } = useEvent();
+    const { displaySnack } = useSnack();
 
+    //-- Subscription --//
     const enqueuedPushConfig = React.useMemo<GraphQLSubscriptionConfig<useOnDeckDequeuedSubscription>>(
         () => ({
             variables: {
@@ -71,60 +66,7 @@ export function useOnDeckDequeued({ connections, topic: currentTopic }: Props) {
 
     useSubscription<useOnDeckDequeuedSubscription>(enqueuedPushConfig);
 
-    // NOTE: Should always order the onDeck list from lowest to highest
-    // That way, whenever the list is empty and a new one is added (which is set to the time in ms),
-    // it will always be after the current question.
-    const calculatePosition = React.useCallback(
-        (list: Question[], movedQuestionIndex: number) => {
-            // If the list is length 1 then it is likely the first item added to the list, calculate a new position
-            if (!list || list.length <= 1) {
-                const currentTimeMs = new Date().getTime();
-                const currentTimeMsStr = currentTimeMs.toString();
-                const calculatedPosition = parseInt(currentTimeMsStr);
-                return calculatedPosition;
-            }
-
-            // The source indx is useless here since we are moving from a different list, can only use destination index
-            // The destination index will be where the moved quesiton is as the list is updated while moving it.
-            // Should check if there at the end of the list or the start of the list
-            // If not, then calculate the position based on the two questions around it
-            // Already handled case with it being the first, so if the index is 0 then there should be at least one question below it
-            if (movedQuestionIndex === 0) {
-                // If the index is 0 then the new position should be less than the next question in the list
-                const nextQuestion = list[movedQuestionIndex + 1];
-                const nextQuestionTopic = nextQuestion?.topics?.find((t) => t.topic === currentTopic);
-                const nextQuestionPosition = !nextQuestionTopic
-                    ? parseInt(nextQuestion.position)
-                    : parseInt(nextQuestionTopic.position);
-                // NOTE: race condition, since we're using time for ordering, then adding 1000 ms (1s) will mean that the order
-                // at the very end may be messed up, but that's okay, the start is what's important
-                return nextQuestionPosition - 1000;
-            }
-
-            // In this case we should have at least one question above it to reference and calculate the new position
-            const prevQuestion = list[movedQuestionIndex - 1];
-            const previousQuestionTopic = prevQuestion?.topics?.find((t) => t.topic === currentTopic);
-            const prevQuestionPosition = !previousQuestionTopic
-                ? parseInt(prevQuestion.position)
-                : parseInt(previousQuestionTopic.position);
-            const nextQuestion = list[movedQuestionIndex + 1];
-            if (!nextQuestion) {
-                // If there is no next question then the new position should be greater than the previous question
-                return prevQuestionPosition + 1000;
-            }
-            // If there is a next question then the new position should be between the previous and next question
-            const nextQuestionTopic = nextQuestion?.topics?.find((t) => t.topic === currentTopic);
-            const nextQuestionPosition = !nextQuestionTopic
-                ? parseInt(nextQuestion.position)
-                : parseInt(nextQuestionTopic.position);
-            // const previousQuestionPosition = !previousQuestionTopic ? parseInt(prevQuestion.position) : parseInt(previousQuestionTopic.position);
-            const position = Math.round(prevQuestionPosition + nextQuestionPosition) / 2;
-            if (position < -1) throw new Error('Invalid position');
-            return position;
-        },
-        [currentTopic]
-    );
-
+    //-- Mutation --//
     type MutationInput = {
         questionId: string;
         eventId: string;
@@ -132,12 +74,12 @@ export function useOnDeckDequeued({ connections, topic: currentTopic }: Props) {
         movedQuestionIndex: number;
     };
 
-    const [commit] = useMutation(USE_ON_DECK_DEQUEUED_MUTATION);
+    const [commit] = useMutation<useOnDeckDequeuedMutation>(USE_ON_DECK_DEQUEUED_MUTATION);
     const removeFromOnDeck = React.useCallback(
         (input: MutationInput) => {
             const { list, movedQuestionIndex } = input;
 
-            const newPosition = calculatePosition(list, movedQuestionIndex);
+            const newPosition = calculateOnDeckDequeuePosition({ list, movedQuestionIndex, currentTopic });
             if (newPosition <= 0) throw new Error('Invalid position');
 
             commit({
@@ -149,9 +91,17 @@ export function useOnDeckDequeued({ connections, topic: currentTopic }: Props) {
                         newPosition: newPosition.toString(),
                     },
                 },
+                onCompleted: (response) => {
+                    if (response.removeQuestionFromOnDeck.isError) {
+                        return displaySnack(response.removeQuestionFromOnDeck.message, { variant: 'error' });
+                    }
+                },
+                onError: (error) => {
+                    displaySnack(error.message, { variant: 'error' });
+                },
             });
         },
-        [calculatePosition, commit, currentTopic]
+        [commit, currentTopic, displaySnack]
     );
 
     return { removeFromOnDeck };
